lucas-commit

+++ database/migrations/2024_03_14_092329_add_foto_to_users_table.php
@@ -11,12 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        if (!Schema::hasTable('users')) {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('imagem',255)->nullable()->after('remember_token');
-        });
-        }
 
+        Schema::table('users', function (Blueprint $table) {
+            if (!Schema::hasColumn('users','imagem')) {
+                $table->string('imagem',255)->nullable()->after('remember_token');
+            }
+        });
     }
 
     /**
@@ -25,7 +25,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('imagem');
+            if (Schema::hasColumn('users', 'imagem')) {
+                $table->dropColumn('imagem');
+            }
         });
     }
 };
